reorder types

+++ src/buffers.jl
@@ -3,8 +3,8 @@ const MPIFloatingPoint = Union{Float32, Float64}
 const MPIComplex = Union{ComplexF32, ComplexF64}
 
 const MPIDatatype = Union{Char,
-                          Int8, UInt8, Int16, UInt16,
-                          Int32, UInt32, Int64, UInt64,
+                          Int8, Int16, Int32, Int64,
+                          UInt8, UInt16, UInt32, UInt64,
                           Float32, Float64, ComplexF32, ComplexF64}
 MPIBuffertype{T} = Union{Ptr{T}, Array{T}, SubArray{T}, Ref{T}}
